requirements.txt
requests
beautifulsoup4
fpdf
googlesearch-python
pdfminer.six
